:recycle: Moved learning paths controller unit test to its appropriate package.
