+++ src/main/java/com/codurance/allaboard/core/model/topic/Subtopic.java
@@ -1,0 +1,4 @@
+package com.codurance.allaboard.core.model.topic;
+
+public class Subtopic {
+}
